rework test/setup.js + add vhtlc.test.ts file
