+++ docker-compose.yml
@@ -24,7 +24,7 @@ services:
         environment:
             - ARKD_ROUND_INTERVAL=10
             - ARKD_NETWORK=regtest
-            - ARKD_LOG_LEVEL=5
+            - ARKD_LOG_LEVEL=6
             - ARKD_VTXO_TREE_EXPIRY=20
             - ARKD_TX_BUILDER_TYPE=covenantless
             - ARKD_ESPLORA_URL=http://chopsticks:3000
